# src/iron-pets/backend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Iron Pets API
  version: 1.0.0
  description: |
    REST API for Iron Pets e-commerce platform - MVP version.
    Provides endpoints for authentication, product catalog, shopping cart, checkout, orders, and user management.
  contact:
    name: Iron Pets Engineering Team
    email: engineering@ironpets.com
  license:
    name: Proprietary
servers:
  - url: https://api.ironpets.com/v1
    description: Production server
  - url: https://staging-api.ironpets.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /user/password:
    put:
      tags: [User]
      summary: Change password
      description: Update user password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/account:
    delete:
      tags: [User]
      summary: Delete account
      description: Permanently delete user account and data
      operationId: deleteAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Cannot delete account with pending orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/addresses:
    get:
      tags: [User]
      summary: List addresses
      description: Get all user saved addresses
      operationId: listAddresses
      responses:
        '200':
          description: Addresses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressListResponse'

    post:
      tags: [User]
      summary: Add address
      description: Add new saved address
      operationId: addAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '201':
          description: Address created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /user/addresses/{id}:
    put:
      tags: [User]
      summary: Update address
      description: Update existing address
      operationId: updateAddress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: Address updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [User]
      summary: Delete address
      description: Remove saved address
      operationId: deleteAddress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Address deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ========================================
  # PET ENDPOINTS
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
        - newPasswordConfirm
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
        newPasswordConfirm:
          type: string
          minLength: 8
    DeleteAccountRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string

    # ========================================
    # ADDRESS SCHEMAS
    # ========================================
    Address:
      type: object
      required:
        - firstName
        - lastName
        - addressLine1
        - city
        - state
        - zipCode
        - phone
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        addressLine1:
          type: string
          maxLength: 255
        addressLine2:
          type: string
          maxLength: 255
          nullable: true
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        zipCode:
          type: string
          maxLength: 20
        phone:
          type: string
          maxLength: 20
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AddressRequest:
      type: object
      required:
        - firstName
        - lastName
        - addressLine1
        - city
        - state
        - zipCode
        - phone
      properties:
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        addressLine1:
          type: string
          maxLength: 255
        addressLine2:
          type: string
          maxLength: 255
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        zipCode:
          type: string
          maxLength: 20
        phone:
          type: string
          maxLength: 20
        isDefault:
          type: boolean
    AddressResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Address'
    AddressListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Address'

    # ========================================
    # PET SCHEMAS
    # ========================================
...